% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory.R
\name{cor_nz}
\alias{cor_nz}
\title{Function to compute correlation based on non-zero values}
\usage{
cor_nz(m, min_exprs = 0, method = "pearson")
}
\description{
Function to compute correlation based on non-zero values
}
