% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq.R
\name{de_heatmap}
\alias{de_heatmap}
\title{Make a heatmap of DE genes}
\usage{
de_heatmap(exprs_matrix, interesting_genes, annotation,
  key_label = "expression level", mar = c(10, 10), cexRow = 1,
  cexCol = 1)
}
\description{
Make a heatmap of DE genes
}
