% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory.R
\name{expression_hist}
\alias{expression_hist}
\title{Make a histogram of gene expression levels}
\usage{
expression_hist(count_matrix, title = "")
}
\description{
Make a histogram of gene expression levels
}
