% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{fetch_DataFrame}
\alias{fetch_DataFrame}
\title{Wrapper function for reading data
from an SQLite databases}
\usage{
fetch_DataFrame(sql_query, database, attach = NULL)
}
\arguments{
\item{sql_query}{An sql query statement}

\item{database}{Location of an sqlite database}
}
\description{
Wrapper function for reading data
from an SQLite databases
}
