% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readqc.R
\name{fetchFastQC}
\alias{fetchFastQC}
\title{A helper function to retrieve Fastqc results from the pipeline_readqc database}
\usage{
fetchFastQC(qc_metric, db = fastqc_sqlite_database)
}
\description{
A helper function to retrieve Fastqc results from the pipeline_readqc database
}
