% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readqc.R
\name{fetchReadqcLegacy}
\alias{fetchReadqcLegacy}
\title{A helper function to retrieve Fastqc results from the pipeline_readqc database}
\usage{
fetchReadqcLegacy(qc_metric, db = readqc_db,
  name_fields = qc_name_field_titles, plot_group_names = plot_groups,
  paired = TRUE)
}
\description{
A helper function to retrieve Fastqc results from the pipeline_readqc database
}
