% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq.R
\name{filterDESeqDatasetByGroup}
\alias{filterDESeqDatasetByGroup}
\title{Group-aware pre-filtering of a DESeq object.
This function is used to filter out genes that do not pass a minimum
expression level in a minimum number of replicates from at least 1 group.}
\usage{
filterDESeqDatasetByGroup(
  dds,
  grouping_var = "condition",
  min_replicates = 2,
  min_counts = 10
)
}
\arguments{
\item{dds}{A DESeq2 dataset object}

\item{grouping_var}{The name of the variable in the colData that contains the experimental groups}

\item{min_replicates}{The minimum number of replicates within a group that have min_counts}

\item{min_counts}{The minimum number counts}
}
\description{
Group-aware pre-filtering of a DESeq object.
This function is used to filter out genes that do not pass a minimum
expression level in a minimum number of replicates from at least 1 group.
}
