% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory.R
\name{get_col}
\alias{get_col}
\title{Function to get colors for a heatmap}
\usage{
get_col(mat, lower_quantile = 0, upper_quantile = 0.9, nbreaks = 50,
  palette = brewer.pal(9, name = "YlOrRd"))
}
\arguments{
\item{mat}{the data matrix}

\item{lower_quantile}{the quantile to use for the lowest break}

\item{upper_quantile}{the quantile to use for the highest break}

\item{nbreaks}{the number of breaks}

\item{palette}{An RColorBrewer palette}
}
\description{
Function to get colors for a heatmap
}
