% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_den}
\alias{get_den}
\title{function to return a cluster dendrogram
with optional optimised leaf order.}
\usage{
get_den(m, optimize = F, dist_method = "manhattan",
  clust_method = "complete", p = 1.5)
}
\description{
function to return a cluster dendrogram
with optional optimised leaf order.
}
