% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getExamples}
\alias{getExamples}
\title{Copy the report template to the current working directory}
\usage{
getExamples(topic = "clustering")
}
\arguments{
\item{topic}{The topic of the examples to fetch}
}
\description{
Copy the report template to the current working directory
}
