% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getReport}
\alias{getReport}
\title{Copy the report template to the current working directory}
\usage{
getReport(report = "readqc")
}
\arguments{
\item{report}{The report template to fetch ("readqc" or "rnaseq")}
}
\description{
Copy the report template to the current working directory
}
