% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{labelClusters}
\alias{labelClusters}
\title{function to draw a heatmap with colored clusters}
\usage{
labelClusters(clust, title, k = 10, min_size = 100)
}
\arguments{
\item{clust}{An object with m=matrix, den_genes=gene dendrogram, den_samples = sample dendrogram}

\item{title}{Title}

\item{k}{The number of clusters}

\item{min_size}{The minimum size}
}
\description{
function to draw a heatmap with colored clusters
}
