% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{parPvclust_opt}
\alias{parPvclust_opt}
\title{Version of parPvclust() with optimised leaf ordering
# e.g. to execute a clustering in parallel
# library(snow)
# cl <- makeCluster(12, type="SOCK")
#
# pvclust.result <- parPvclust(cl, data.matrix,
#                           method.dist="correlation",
#                           method.hclust="average",
#                           nboot=1000)
# stopCluster(cl)}
\usage{
parPvclust_opt(
  cl = NULL,
  data,
  method.hclust = "average",
  method.dist = "correlation",
  use.cor = "pairwise.complete.obs",
  nboot = 1000,
  r = seq(0.5, 1.4, by = 0.1),
  store = FALSE,
  weight = FALSE,
  init.rand = NULL,
  iseed = NULL,
  quiet = FALSE
)
}
\arguments{
\item{cl}{Snow cluster}

\item{data}{The data matrix}

\item{method.hclust}{The hclust method}

\item{method.dist}{The distance method}

\item{use.cor}{For computing correlation distance}

\item{nboot}{The number of bootstraps (for pvclust)}

\item{r}{A sequence.}

\item{store}{pvclust arg}

\item{weight}{pvclust arg}

\item{init.rand}{pvclust arg}

\item{iseed}{pvclust arg}

\item{quiet}{pvclust arg}
}
\description{
Version of parPvclust() with optimised leaf ordering
# e.g. to execute a clustering in parallel
# library(snow)
# cl <- makeCluster(12, type="SOCK")
#
# pvclust.result <- parPvclust(cl, data.matrix,
#                           method.dist="correlation",
#                           method.hclust="average",
#                           nboot=1000)
# stopCluster(cl)
}
