% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{plotDensities}
\alias{plotDensities}
\title{Helper function for plotting densities}
\usage{
plotDensities(countMatrix, phenoData, plot_color, plotTitle)
}
\description{
Helper function for plotting densities
}
