% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq.R
\name{run_deseq2}
\alias{run_deseq2}
\title{Function for running a DESeq2 test.}
\usage{
run_deseq2(dds, name = "placeholder", location_function = "median",
  fit_type = "local", contrast, id_col = "gene_id", annotation = ann_df)
}
\description{
Function for running a DESeq2 test.
}
