% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq.R
\name{setup_dds}
\alias{setup_dds}
\title{Function for preparing a DESeq2 object}
\usage{
setup_dds(counts, columnData, subset = NULL, design = NULL,
  reference_levels = NULL)
}
\description{
Function for preparing a DESeq2 object
}
