% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{upperQuartileNormalise}
\alias{upperQuartileNormalise}
\title{Upper-quartile normalise an expression matrix}
\usage{
upperQuartileNormalise(exprs_matrix)
}
\arguments{
\item{exprs_matrix}{An expression matrix}
}
\description{
Upper-quartile normalise an expression matrix
}
