% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{write_DataFrame}
\alias{write_DataFrame}
\title{Wrapper function for writing data
to an SQLite database}
\usage{
write_DataFrame(TableName, DataFrame, database, overwrite = FALSE,
  rownames = FALSE)
}
\arguments{
\item{database}{Location of an sqlite database}

\item{sql_query}{An sql query statement}
}
\description{
Wrapper function for writing data
to an SQLite database
}
