% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq.R
\name{get_interesting}
\alias{get_interesting}
\title{Get interesting genes...}
\usage{
get_interesting(tt, p_value_threshold = 0.05, n = 10, id_col = "gene_id",
  p_col = "padj", fc_col = "log2FoldChange")
}
\description{
Get interesting genes...
}
